Fix field visibility issue in LocalDatabase
It tracks current availability requirement in a field. This field is mostly
accessed from the constructor and lifecycle methods `#start()` and `#stop()`.
Constructor assignment is safe and lifecycle methods are synchronized.

Field is also accessed from `#isAvailable()` method which is not synchronized
and thus it's access was racy. This commit makes field volatile to ensure
correct visibility in `#isAvailable()`.