Check point scheduler cause panic after a couple of failures
This is in addition to the fact that check pointer logs failures.
The background job check pointing when needed can fail a couple of times
before treating that as a panic, but will eventually do if check pointing
continues to fail. Previously this wouldn't happen, in fact check pointing
job would stop entirely after first failure without mentioning it or
causing panic.