recover skips in segmented raft log
This piece of functionality was missing and could lead to the log claiming
to have entries which it does not, leading to liveness issues in the
core Raft logic/interactions where an entry could never be served but
would continually be attempted to. This was observed in ITs, very rarely,
and required an additional set of circumstances to cause the actual liveness
issue. Afaik this never compromised data safety because the entries were simply
never served and the interaction was stuck in an infinite loop.

The fix consists of detecting skips by checking the headers of the segments. A
skip is identified by seeing that there is a gap between the previous segments
last index and the current files prevIndex. Note that both of these values are
stored in every header, i.e. the current header keeps the previous files last
index since that is written down on creation (due to rotation or skip).