Failed counts store rotation leaves it in good state
Previously a rotation failure would leave the counts store in a state of perpetual
rotating state as well as having the store file closed, i.e. a rather broken state.

With this change a failed rotation will revert back to the previous state and just
continue working as it would before rotation was attempted.