Use ReadWriteLock for log rotation
Changes the locking to use ReadWriteLock instead of synchronized
blocks. This should help to increase throughput on the logging while
keeping the required exclusive locks on Closing and Rotation.

Changed OutRef to be as a simple Object instead of a AtomicReference
since we do locking and wrap the object.

Removed all references to rotated log files. There should be no need
to keep holding on to these since we close the log file before rotation.