Sketch of index population that only loads relevant properties
The new NodeUpdates object initially gets all labels for the node, before and after, and
information about the property values that are added, removed or changed. The forIndexes
method then analyses all index descriptors and finds what properties that need loading to
determine what indexes need updating. Using the PropertyLoader it loads these properties,
 and returns a sequence of IndexEntryUpdates.

For online index update the glue that gives the NodeUpdates a real PropertyLoader is still missing.