Initial implementation of composite indexes in TxState
The TxState keeps a diffset for every index that has changed, and
thus keeps track of what nodes that have gotten or lost a certain
property value combination that matches the index. This makes
queries straight forward. The complex logic of deciding what indexes
that are affected by a certain label or property change is done on
write, in the new class IndexTxStateUpdater.