Fix race bug in page faulting
If a thread observes a missing entry in the translation table, it will attempt to initiate a page fault.
Only one thread at a time should try to fault any given page, and this is guarded by the LatchMap.
However, it can be the case that another thread completes the fault in between us noticing that a fault is needed, and trying to grab a latch.
Since the other thread completed the page fault, and released its latch, we will get our own latch and mistakenly do another page fault.
We now guard for this by double-checking that the entry is indeed still missing after we have gotten our latch.
If the entry is no longer missing, then it means that another thread completed the fault ahead of us, and so we simply release the latch we got and retry our pin.