Fix ReadPageCursor shouldRetry bug with linked cursors
When we checked the lock validity using the stamp from the top-most page cursor, instead of the stamp for the given cursor.
This obviously often resulted in the check failing, with no way to get out of the shouldRetry look.
The shouldRetry check now currectly uses the lock stamp for the given cursor in the chain, and also makes sure to refresh the lock state of all the cursors in the chain when a restart is necessary.