Fixes issues causing import tool to sometimes hang when panicing
The nature of this problem was racy and it could result in the whole import
tool seemingly hang, not even printing the cause of the panic.

The issue have been fixed mostly by simplifying code and unifying concepts
around panic so that such is treated the same by all import participants.