Update component injection validation rules
@​Context field types can now be categorized in three ways:

 1. supported: it is legal for any procedure|function. It can be:
    - Log
    - TerminationGuard
    - GraphDatabaseService
    - SecurityContext
    - ProcedureTransaction
 2. restricted: it can be used if and only if the procedure
 |function is declared in `dbms.security.procedures.unrestricted`
 setting. The type can be:
    - GraphDatabaseAPI
    - KernelTransaction
    - DependencyResolver
    - UserManager
    - EnterpriseAuthManager
    - SecurityLog
 3. unsupported: it is illegal (any other type, basically).

The processor cannot know whether a procedure|function is
unrestricted at the time of compilation and will, by default, emit
a warning if such a type is encountered. The existing processor
option `-AIgnoreContextWarnings` can be used to mute this.

From now on, if a type is unsupported, a compilation error will be
raised.