Ability to auto-select label index based on existing store
Introducing LabelIndex.AUTO which is intended to be used for tooling
where it's convenient to not have a say in which label index to use,
but let database select which ever label index is present already.

This was driven by a backup scenario where the backup client would
select a different label index than the source. This would work
conceptually (because it would have been rebuilt), but proved to be
troublesome by means of page cache actions for moving files after
temporary store had been started and stopped. In this case the
original label index store would have been deleted and making the
following move action invalid.

Letting the backup client internally use AUTO label index selection
solves this. It's still of course possible to configure another
label index when restoring from this backup, which is separate from
this problem of doing the backup.