Adds PositionableChannel interface to be able to set channel position
Only some channel implementations can implement this interface, e.g. ReadAheadChannel
and InMemoryClosableChannel. InMemoryClosableChannel had to be split into reader/writer
because there was confusion about which of the read/writer set/get current position
would otherwise apply to. Its surface still looks the same with the optional ability
to get a reader or writer which is either or.