Overwriting property with same value no longer generates write command
It seems obvious, but setting a node/relationship property to the same value
as it already is doesn't require any change to that node/relationship.
Even so this was the case previously. A command would be generated, property
record would be updates on apply, indexing changes would be applied...
all for nothing.