Cluster has ability to recover from half-applied constraint
Creating a constraint involves two transactions:
- one that creates the backing constraint index, where index population
  starts once committed and applied
- one that creates and activates the constraint after successful index population

There was a problem in distributed environments where this could happen:
- constraint would begin being created on a cluster instance A, i.e.
  the first transaction creating the constraint index
- another cluster instance B would pull and apply this transaction
- B would crash
- A would complete the constraint creation by commiting the second
  transaction activating the constraint
- B would be started again, where recovery would treat this half-baked
  constraint as an unsuccessful constraint creation and mark it as failed.
  (single mode differs here where the constraint index would even be deleted)
- when B would continue to pull transactions from A and eventually get to
  the transaction activating the constraint, B would not be able to apply it
  because the backing index had been marked as failed during recovery

Fix proposed here is to not try and be helpful and delete these "orphaned"
constraint indexes during startup because even if that only happened in
single db mode then an instance may switch back and forth between
clustered and single, making this hard to look consistent to user.

The other part of the fix is to remove the optimization to not wait
and immediately mark such constraint indexes as failed during recovery,
now that they will not be deleted.

Additionally instead of cleaning up orphaned constraint indexes, such
indexes can be reused for another constraint. So a retry of such a transaction
after a crash would work, just as it would before this change.
Having the ability to reuse orphaned constraint indexes can have multiple
concurrent transactions wanting to reuse it to fight over it. One of them will
win and the others will fail, this is concurrently safe because of
the schema write lock when creating the second transaction that
activates the constraint.