Move loading of additional properties back in OnlineIndexUpdates
When determining if a node update is relevant to some index, we might need to know
the existince and values of properties that have not been changed. This was done by
loading all properties on label changes in 3.1, and handled by OnlineIndexUpdates. With the
introduction of composite indexes in 3.2, this behaviour was changes to only load properties
that are known to be relevant to at least one index. The natural place to do this was believed
to be the IndexingService.

This might cause store corruption though, if the properties or labels of the node changes
in-between when the initial NodeUpdates object is created in OnlindeIndexUpdates, and when
the IndexingService actually converts the NodeUpdates into IndexEntryUpdates. This commit
changes the loading of additional properties to once more happen in the OnlineIndexUpdates
class, which should fix these store corruption issues.