GBPTree never checkpoint on close
NativeLabelScanStore take over responsibility for checkpointing
underlying GBPTree. It will do so only if, at shutdown, no panic
has occurred and NativeLabelScanStore was successfully started,
meaning any recovery was completed.

When recovering the tree we rely on the last successful checkpoint
before the crash that made recovery necessary. If recovery fails we
will retry it again on next start up. If the tree was checkpointed
then the next recovery will not be able to reliably clean up all
'crashed' pointers.