Never checkpoint GBPTree on close
Checkpoint will increment stable and unstable generation of tree,
making it impossible to recognise crash-pointers during the recovery
following the database crash.

Responsibility for checkpointing is completely left to the system
so GBPTree will only checkpoint when rest of system is checkpointing.

First attempt at this prevented checkpoint on close only when system
saw panic, but that did not guard for some failures during recovery
that did not trigger panic.