Asynchronous crash pointer cleanup
Before, cleanup of crashed pointers was blocking
and was done as part of GBPTree constructor.

Now, GBPTree just register a cleanup job to
provided RecoveryCleanupWorkCollector that will
be triggered before next checkpoint and cleanup
will be performed asynchronously by some background
thread.

It is safe to have crash pointer cleanup work
concurrently with GBPTree writer because:
- Writer will treat all crashed pointers exactly
  like empty pointers. Ignore when reading and
  overwrite when writing.
- Crash pointer cleaner will overwrite all
  crashed pointers with zeroes, effectively
  turning them into empty pointes.
- Crash pointer execution is not at all affected
  by content of the tree. Just the crashed pointers
  so concurrent structure modifications does not
  matter.