Don't force flush as part of recovery
Flushing will implicitly perform a checkpoint on GBPTree.
This will close the generation gap that is necessary to
recognize CRASH pointers in the tree and thus asynchronous
crash gen clean job will not be able to do it's job.

Flushing was the way checkpoint was done before explicit
checkpoint became a thing. And flushing in recovery is an
artifact from that time, it does not do anything useful
other than making recovery take longer time.