Only check uniqueness constraints on actual changes
Skip checking if a property is re-assigned to the same value.
Skip checking if a label is added that already was assigned.

This avoids a deadlock scenario between concurrent MERGE queries, where the
MERGE was followed by re-assignment of the same property. This is quite a
common pattern, especially to bulk-re-assign all expected properties of the node.