New page lock transition: write->flush
It's now possible to atomically undo a write lock, and attempt to grab a flush
lock, on a page in PageList.
This will be useful for adding a PF flag that instructs write page cursors to
flush the page on unpin.
This in turn will be useful for certain stages or phases in the parallel batch
importer, where a group of reader threads are reading in pages ahead of a group
of writer threads who then modify those pages. Currently the reads are fielding
most of the page faulting, and thus also end up doing most of the flushing.
That is, the reads end up doing both the read and the write IO.