Better error message for existing index
If you first created a composite index and then tried to create a node
key constraint on the same properties we failed when trying to create the
proper error message. Instead we got the error message:

```
Single property schema requires one property but had 2
```
which was caused by calling `getProperty` when creating the error message.