Bump unstable generation in GBPTree constructor
When starting up on an existing file, all changes that
was flushed to disk between last checkpoint and start will be
seen as 'crashed' whether or not the tree was properly closed.

The only way to see a clean state when opening a tree on existing
file is if checkpoint was called prior to close and no changes
happened in between.