GBPTree internally handle cleaning of crashed pointers
By checking if last close was "clean" (that there where no changes since last
checkpoint) the GBPTree knows if cleaning is necessary or not. If tree is dirty a
clean will be performed as part of construction.

Adding clean as part of tree state means bumping 'format'. Hence the new
'current-format.zip'.

GBPTree no longer needs to be told from outside to prepare for or finish recovery.

Remove readOnly mode for GBPTree and let that be handled fully from the outside.
Why? Because even if NativeLabelScanStore is started as read only it will still need
to be able to create a complete new instance of underlying tree and populate it by
scanning node store.