Lost properties in property existence enforcing
Fix error on arrays union that was causing result array missing part of the values.
One of the observable effects was that property constraint enforcer missed part of the properties
and it was possible to create nodes with properties set that do not satisfy all defined constraints.

Add missing tests for node key constraint creation and validation.
Move ImpermanentEnterpriseDatabaseRule to correct package.