Input filtering in importer still shows progress
Previously filtering of input data could potentially have large sections
of input be filtered and so now progress displayed. Since both progress
and processor assignments is based on number of processed batches the
behaviour of the iterator batch now changed so that it sends empty no-op
batches for the skipped entries.