Floating point comparisons by scaling up precision
When comparing a JVM float with a JVM double, indexes behaved
differently from the runtime comparison. The runtime comparison opted
for scaling down the precision to compare between two floats,
whereas the indexes scaled up precision.

This change makes sure to have the same semantics no matter if an index
exists or not.