Fixes race between starting transaction and shutdown
This was designed to work already, but there was a slight gap in that
armor. DatabaseAvailability is the among last components to register in
the database lifecycle and therefore one of the first to be stopped.
Stopping DatabaseAvailability will raise the availability guard and then
await all active transactions to be finished (with some configurable
timeout). Starting transactions also checked the availability guard
state and registered themselves as active.

The problem was that the active state wasn't registered until after the
availability guard had been checked and so this little window between
those calls was racing with that same window in DatabaseAvailability
potentially allowing a transaction to start without being aware that
database was shutting down and shutdown process being unaware of an
active transaction running concurrently. This could result in
unspecified exceptions being thrown due to transaction running inside
components which were shutting down or even had been shut down.

This commit tightens this gap by letting transaction starting register
itself as active before checking availability guard. This way
transactions starting concurrently with shutdown will deterministically
succeed or fail, all depending on whether they made it in before the
availability guard was raised for shutdown.

This still leaves the max timeout that DatabaseAvailability waits before
going ahead with the shut down anyway. But that is by default a timeout
reasonable for most setups and it's configurable.