Rewrite of how data from Input gets written into records
Instead of going through the stages w/ the queuing and passing objects
importing is done by fully executing the importing threads in parallel.
All data structures and access patterns ihave been made to allow for
concurrent access.

Garbage was a big concern previously and so this new importing doesn't
create objects for the entities, but instead allow visitors to observe
the different fields, directly from an underlying buffer which has been
read from the input source.