changed default transaction log retention to 1 day in community edition
This was probably overaggressive for community users. Which then leads
to public slack conversations about why is the size of my graph.db XXX
GB when I've only loaded X gig of data. And then how/when/why can I
clean out all the neostore.transaction* files. For community members,
for which neither HA or backup functionality is used and for which
this functionality requires said logs, keeping 7 days is
overaggressive.