Make SessionResetIT less flaky
Class contains tests that verify `Session#reset()` java driver
functionality in Bolt server. They were quite flaky.

Commit fixes CSV file URI on Windows and adds closing of the driver in
the randomized `#reset()` test. Later is to make sure assertions are
made on a database that is not talking to any driver.