Awaits current batches to complete before changing processor count
in ForkedProcessorStep, so that there may not be an order inconsistency.
The order inconsistency would come from there being a different
processor count and so the distribution of work would be different,
which means that a new particular path of a batch may be assigned to
a different processor than the one getting those parts, before the
processor change. This would introduce a race between the new and old
processor for that particular part.

Solution here is to drain the processing queue before applying the new
processor count.