Fix page cursor usage bug in GBPTree
It was possible for an inconsistent read of a read page cursor to cause an out
of bounds access to a write page cursor, when writing the page header during
state forcing.
The fix is to explicitly clear the bounds flag on the write cursor, for every
shouldRetry iteration of the read cursor.