Fix a rare bug in GBPTree shutdown
If the page cache eviction thread encounters an exception during eviction, such
as running out of storage space, then that exception will be stored in a field
an delivered to the next unfortunate passersby accessing the page cache.
If this is the GBPTree in the process of shutting down, then the exception
would bubble out without any attempt at recovering the situation.
This behaviour was different from how the record stores shut down, and it would
prevent shut down even after the fault had been corrected, such as more space
becoming available.