Don't let vacuum cause evictions unless there is a pressing need
Previously, we would try to vacuum as soon as 20 files had been unmapped. Such
activity is common around the startup of the database, when the page cache has
plenty of free pages. This caused unexpected evictions to occur quite early in
the life of a database, which is technically fine, but caused many tests to
fail.

Now we don't do that; if the eviction thread has not already started, then we
won't vacuum until there is a "pressing need" in the sense that there are less
than 200 swapper ids left.