Fix bug in owning pipe registration
This task was propagated through the expression tree using rewrite,
which is nut suitable as rewrite performs immutable copy and modify,
while the owning pipe is implemented as mutable state in the expression.
Even though the arguably best solution might be to make the owning pipe
state immutable too, that was considered to be too invasive for a
maintenance branch. Therefor the current solution keeps mutable owning
pipe state, but registers using visit() instead.