More debug info for failed transaction apply
We're not in a good spot if we fail to apply a transaction. Such failures would
happen after a transaction has already been committed, or during recovery. Both
are awful places for exceptions to pop out.

This commit captures exceptions that occur during apply, and wraps them in
another exception that adds information about the transaction, e.g. log
position, which might be helpful for debugging the cause of the failure.