Include a root-cause message in the LifecycleException message
We see that time and time again, people don't follow the "Please see the
attached cause exception" instructions, and that we have to ask them to go back
to their logs and extract the _complete_ stack trace.

This commit won't solve that underlying problem, but now we will at least have
the message for the root cause in the initial report, which is a little bit
more to work with while we wait for people to return with the complete stack
trace.