IndexPopulator for native schema number index
Implemented using GBPTree and for the time being only IndexPopulator is implemented,
online IndexAccessor and SchemaIndexProvider will come later.

Both unique and non-unique layouts are supported and all types of Numbers. There are small
differences between unique and non-unique layouts where entity ID is in the key for
non-unique, but in the value for unique. This makes the keys smaller for the unique layout,
something which will make the overall index size smaller due to internal tree nodes
not containing values. Each layout contains both the comparison value, a coerced value type
which all numbers are converted into for comparison, but also their original value.
This means that there will not need to be an additional entity property lookup for filtering
those results higher up in the stack.