Record use of fixed reference format records in the transaction log
In Neo4j 3.0.6 we introduced the fixed-reference variant of the high_limit
format. The fixed references are faster to read and write, but has a much lower
limit on the IDs it can use. Never the less, since all the references are of
a known fixed length, it can omit the meta-data for reference width that the
variable reference format cannot. Thus, it is sometimes possible for the fixed
reference format to store some records in a single record unit, where the
variablce reference width format would use two record units.

It turns out that we did not store in the transaction log, whether a record
would be applied to the store using a variable width, or a fixed width,
reference format. The absence of this information gets interpreted as if we are
always using the variable length reference format. This could cause problems
during recovery or cluster store copy. A record that fits a record unit using
the fixed reference format, but would require two record units using the
variable width format, would end up overwriting a part of its adjacent record,
or would go out of bounds on the page.

To solve this problem, we now record to the log whether a record was using the
fixed width, or the variable width record format. Fortunately, the record
headers in the transaction log had plenty of spare bits for this purpose. The
previously unused bit that I have taken for this purpose is always zero in the
existing transaction logs. This mirrors the bug exactly, in that the use of
fixed reference format defaulted to `false`. Thus, it can be argued that taking
this bit into use is not a format change. For this reason, I introduce no new
log version, _and_ I fix the bug for both 3.0.x log formats.