Bump transaction log version
With the fix for storing the use of fixed-reference width format in the
transaction log, we have to bump the transaction log version in order to
prevent mixed-version clusters.

In a mixed-version cluster, an old version would be able to produce and
serialise commands in the fixed-width reference format with the "use fixed
reference format" bit lowered. When transferred to newer members of the cluster
where the bug is supposed to be fixed, it would reappear.

With a new transaction log version, this cannot happen.