Unify unique and non-unique index key and value
The only difference between unique key:value and non-unique
key:value was that entityId was stored in key for non-unique
and in value for unique.
The only positive thing about this was that we saved a tiny bit
of storage in the unique case, under 0.1%. The trade off was two different
class hierarchies for key, value and layout and lot's of complicated
generic types in index populator and tests.

Now those hierarchies are collapsed as much as possible.
Note especially that the ValueMerger now need both KEY
and VALUE to correctly report conflicting updates.