Stores actual value in NumberKey
This to be able to handle cases where two longs coerce to the same double value
and to handle that comparison needs to be made on actual values in the key.
This actually opened up an opportunity for reducing the entry size by only storing
the actual value and implementing comparison for those directly.

For the future: if the comparison logic on the actual values are too slow then
the key will have to include the compareValue too, inflating the entry size again.