Consistency checker now check label scan store for missing entries
Previously, it was only verified that everything in label scan store
was consistent with the real store. It was never checked that
everything in real store was actually in the label scan store.

When iterating over the label scan store, we now fill the gaps
by associating every node id that is not present in the label
scan store with an empty long[] (representing the node has no
labels according to the label scan store). This is considered to be
consistent with the node having no labels or not being in use at all.