PBI: Speed up thread coordination in ForketProcessorStep
We now allow a queue of incoming batches of over 200 batches. This reduces the back and forth between the forked processor step and the upstream step, because they no longer need to make progress in lock-step.
The incoming submission CAS-loop has been replaced with `getAndSet`, which has the same effect but cannot fail, making this part of the batch submission wait-free.
The CompletedBatchSender thread now has a proper name, so we can tell what the thread is for in thread listings and thread dumps.