Add a `WorkSync.applyAsync` method for asynchronous application of work
The asynchronous work might be done in the submitting thread, or in any other
thread that interacts with the given `WorkSync`.

The submitting of asynchronous work will never block. All blocking is deferred
to the `AsyncApply.await` method. This method must be called before the work
can be expected to have been applied, and then only if it does not throw an
exception.

Additionaly, if the work ends up being applied in a different thread, then any
thrown exception will only propagate to that thread. The `AsyncApply`
insteances will still be marked as done, but there will be no way to tell if
the application was successful or exceptional.

If the work submitters require such information to be propagated, then they
must encode such propagation infrastructure into the work itself.