Unflake LonelyProcessingStepTest
When a panic is raised in a step, it is immediately marked as completed.
This meant that the isCompleted loop could in rare cases pass through before endOfUpstream had been called by the processor thread.
When this happened, the test would fail. To fix it, we instead wait for the endOfUpstream method to release a latch.
In turns out that this also makes the test faster. Wins all around.