Unflake IndexingServiceTest
There were Mockito.verify calls accepting timeouts, which on some machines would
run out before the condition was triggered. Solution here is to explicitly wait
for the condition before and just verify afterwards. Problem is that the way the
test is set up there's a race between the populator and verifier because an
IndexProxy is marked as failed (in flip) before actually calling
IndexPopulator#close(boolean), so the verify call still has a timeout, although
now the time between those two conditions should be microseconds.