Mark UnavailableException with the correct status code
The intention is for driver software to forget about this server
temporarily when unavailable (until next TTL tick).

Effectively this will mean that a server doing large operations
like a store copy or even shutting down will remove it from
the current set.

The specific status code was already used in the TransactionIdTracker
as well but it should only trigger there in uncommon cases where such
a server is hopelessly behind.