Noop lock manager for read replicas
It throws when you try to take exclusive locks to catch
misusage early. This should generally be writing queries
and if they are aimed at a read replica then we want to
catch and throw early.

If someone uses an exclusive lock in a reading query
then they should fix that and move over to use a shared
lock instead.